requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
httpie==3.2.2
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.2
pytest-django==4.8.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
